+++ middleware/auth.js
@@ -3,7 +3,7 @@ const jwt = require('jsonwebtoken');
 module.exports = function(req, res, next) {
 	const token = req.header('auth-token');
 	if (!token) {
-		return res.status(401).send('Access Denied!');
+		return res.status(401).send('You need to be logged in');
 	}
 
 	try {
