+++ routes/events.js
@@ -1,6 +1,7 @@
 // Route for Events Page
 const express = require('express');
 const router = express.Router();
+const authorize = require('../middleware/auth');
 const EventsModel = require('../models/EventsModel');
 
 // Get All the Events
@@ -11,6 +12,34 @@ router.get('/', async (req, res) => {
 	} catch (err) {
 		res.json({ message: err });
 	}
+});
+
+// Go To Create An Event Page
+router.get('/create-event', (req, res) => {
+	res.send('<h1>Create Events Page</h1>');
+});
+
+// Create An Event POST
+router.post('/create-event', authorize, (req, res) => {
+	const formData = {
+		name: req.body.name,
+		description: req.body.description,
+		address: req.body.address,
+		eventImg: req.body.eventImg,
+		eventDate: req.body.eventDate,
+		authorId: req.user
+	};
+
+	const newEventsModel = new EventsModel(formData);
+
+	newEventsModel
+		.save()
+		.then((dbDocument) => {
+			res.status(201).send(dbDocument);
+		})
+		.catch((error) => {
+			res.status(409).json({ message: error.message });
+		});
 });
 
 // Get a Specific Event
