+++ routes/users.js
@@ -12,7 +12,7 @@ router.post('/singup', (req, res) => {
 	// LETS VALIDATE THE DATA BEFORE WE SUBMIT A USER
 	const { error } = signupValidation(req.body);
 	if (error) {
-		return res.status(400).send(error.details[0].message);
+		return res.status(400).send(`${error.details[0].message} validation problem`);
 	}
 
 	// 1) Capture user account details (e.g first name, last name, etc.)
